Renames SessionForm page into AddSession and creates two new components that will be displayed within the AddSession page, the SessionBreakdown and the actual SessionForm
